Upgrade package version and move to compiled preset
