+++ package.json
@@ -63,7 +63,7 @@
     "@vitejs/plugin-react": "^3.1.0",
     "auto": "^10.3.0",
     "boxen": "^5.0.1",
-    "chromatic": "6.19.0-canary.3",
+    "chromatic": "^6.19.1",
     "dedent": "^0.7.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
