+++ tsup.config.ts
@@ -1,18 +1,31 @@
 import { defineConfig } from "tsup";
 
-export default defineConfig((options) => ({
-  entry: ["src/index.ts", "src/manager.ts"],
-  splitting: false,
-  minify: !options.watch,
-  format: ["cjs", "esm"],
-  dts: {
-    resolve: true,
+export default defineConfig((options) => [
+  {
+    entry: ["src/index.ts"],
+    splitting: false,
+    minify: !options.watch,
+    format: ["cjs"],
+    dts: false,
+    treeshake: true,
+    sourcemap: true,
+    clean: true,
+    platform: "node",
   },
-  treeshake: true,
-  sourcemap: true,
-  clean: true,
-  platform: "browser",
-  esbuildOptions(options) {
-    options.conditions = ["module"];
+  {
+    entry: ["src/manager.ts"],
+    splitting: false,
+    minify: !options.watch,
+    format: ["esm"],
+    dts: {
+      resolve: true,
+    },
+    treeshake: true,
+    sourcemap: true,
+    clean: true,
+    platform: "browser",
+    esbuildOptions(options) {
+      options.conditions = ["module"];
+    },
   },
-}));
+]);
